+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.54.0-beta.0] - Unreleased
 
 ### Added
-- Functions in `qmi.core.rpc` that enable RPC messages to adjust class constants marked as `_rpc_constants` in QMI RPC objects from their proxies.
-- In `qmi.core.util` added functions for checking that two values are of same type, or size and type if the type is an iterable (excluding size for strings). This is used for checking that we set only same type of values (and of same size for iterables) as modifiable `_rpc_constants`.
+- Functions in `qmi.core.rpc` that enable RPC messages to adjust class properties marked as `_rpc_properties` in QMI RPC objects from their proxies.
+- In `qmi.core.util` added functions for checking that two values are of same type, or size and type if the type is an iterable (excluding size for strings). This is used for checking that we set only same type of values (and of same size for iterables) as modifiable `_rpc_properties`.
 
 ### Changed
-- Typed `_rpc_constants` to be a _set_ and changed all definitions to be _sets_.
+- Typed `_rpc_properties` to be a _set_ and changed all definitions to be _sets_.
 
 ## [0.53.0] - 2026-05-11